bump package readmes as part of the "bump" step
